pyproject.toml
[project]
name = "voicecab-flight-booking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "loguru>=0.7.3",
    "pipecat-ai[daily,elevenlabs,google,openai,silero,soundfile]>=0.0.56",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pip>=25.0",
    "poetry>=2.0.1",
    "ruff>=0.9.5",
]
